+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Cleaner functional test helpers
 * Bump HeimGuard
 * Update `UserPolicyHandler` to include a `HasPermission()` implementation
+* Json attribute refactor
 
 ### Fixes